A data entry form allowing people to be entered directly into the person table without needing a drupal account. An example use might be for people wishing to register interest in events without needing a formal account.
Note that this form is a new development, so while it is based on dynamic_location, it is not quite as mature in terms of its functionality. To save current development time, some functionality from dynamic_location is missing from the dynamic_person form currently. There are notes on this in the code, 

1. There is no grid mode.
2. Quite a lot of fields on the Edit Tab inherited from dynamic.php are currently redundant.
3. Person delete functionality is currently commented out, as I am not sure the knock-on effects of this, as people can be linked to user accounts. Enable this when needed and there is time to test.

Most of the functionality is present, including placing fields such as [email address] on the page, and attributes such as [psnAttr:5] and page reloading to edit.